\name{MN-package}
\alias{MN-package}
\docType{package}
\title{
Matrix Normal Distribution
}

\description{
Density computation, random matrix generation and maximum likelihood estimation of the matrix normal distribution. For references see: Pocuca N., Gallaugher M. P., Clark K. M. & McNicholas P. D. (2019). Assessing and Visualizing Matrix Variate Normality. arXiv:1910.02859 and the relevant wikipedia page.

}

\details{
\tabular{ll}{
Package: \tab MN\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2024-05-21 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr} and Omar Alzeley \email{oazeley@uqu.edu.sa}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Pocuca, N., Gallaugher, M. P., Clark, K. M., & McNicholas, P. D. (2019). Assessing and Visualizing Matrix Variate Normality. arXiv:1910.02859.
}





