\name{Regression with a matrix of covariates}
\alias{mn.reg2}
\title{
Regression with a matrix of covariates
}

\description{
Regression with a matrix of covariates.
}

\usage{
mn.reg2(Y, x, tol = 1e-6, maxiter = 100, xnew = NULL)
}

\arguments{
\item{Y}{
A list with k elements (k is the sample size), k matrices of dimension \eqn{n \times p} each.
}
\item{x}{
A matrix with covariates, where the i-th row represents the covariate vector is related
to the i-th response matrix Y.
}
\item{tol}{
The tolerance value to terminate the iterative algorithm.
}
\item{maxiter}{
The maximum number of iterations allowed.
}
\item{xnew}{
If you have new x values and want to predict the values of the response matrices pass them here.
}
}

\details{
The function performs matrix normal regression with a matrix of covariates.
}

\value{
A list including:
\item{runtime}{
The runtime required for the whole fitting procedure.
}
\item{iters}{
The number of iterations required for the estimation of the U and V matrices.
}
\item{loglik}{
The log-likelihood value.
}
\item{B}{
The estimated matrix of coefficients.
}
\item{U}{
The estimated covariance matrix associated with the rows, a numerical matrix of dimensions \eqn{n \times n}.
}
\item{V}{
The estimated covariance matrix associated with the columns, a numerical matrix of dimensions \eqn{p \times p}.
}
\item{est}{
The estimated matrix responses for the new x values, otherwise this is NULL.
}
}

\references{
https://en.wikipedia.org/wiki/Matrix_normal_distribution#Definition

Pocuca N., Gallaugher M. P., Clark K. M. & McNicholas P. D. (2019). Assessing and Visualizing Matrix Variate Normality. arXiv:1910.02859.
}

\author{
Michail Tsagris.

R implementation and documentation: Michail Tsagris \email{mtsagris@uoc.gr}.
}

%\note{
%%  ~~further notes~~
%}

\seealso{
\code{\link{dmn}, \link{rmn}, \link{ddplot}
}
}

\examples{
M <- as.matrix(iris[1:8, 1:4])
U <- cov( matrix( rnorm(100 * 8), ncol = 8 ) )
V <- cov( iris[1:50, 1:4] )
Y <- rmn(150, M, U, V)
x <- iris[, 1:2]
colnames(x) <- c("sepal-length", "sepal-width")
mod <- mn.reg2(Y, x)
}
